refactor: global analytics loader
