+++ app/layout.tsx
@@ -2,9 +2,9 @@
 
 import "./globals.css";
 import { ReactNode } from "react";
-import Script from "next/script";
 import Navbar from "@/components/Navbar";
 import Footer from "@/components/Footer";
+import AnalyticsLoader from "./components/AnalyticsLoader";
 
 export const metadata = {
   metadataBase: new URL("https://gearizen.com"),
@@ -75,34 +75,7 @@ export default function RootLayout({ children }: { children: ReactNode }) {
       suppressHydrationWarning
     >
       <head>
-        <link
-          rel="preconnect"
-          href="https://pagead2.googlesyndication.com"
-          crossOrigin="anonymous"
-        />
-        <link
-          rel="preconnect"
-          href="https://www.googletagmanager.com"
-          crossOrigin="anonymous"
-        />
-        <Script
-          src="https://pagead2.googlesyndication.com/pagead/js/adsbygoogle.js?client=ca-pub-2108375251131552"
-          strategy="afterInteractive"
-          crossOrigin="anonymous"
-          onError={(e) => console.error("AdSense load error:", e)}
-        />
-        <Script
-          src="https://www.googletagmanager.com/gtag/js?id=G-V74SWZ9H8B"
-          strategy="afterInteractive"
-        />
-        <Script id="gtag-init" strategy="afterInteractive">
-          {`
-            window.dataLayer = window.dataLayer || [];
-            function gtag(){dataLayer.push(arguments);}
-            gtag('js', new Date());
-            gtag('config', 'G-V74SWZ9H8B', { anonymize_ip: true });
-          `}
-        </Script>
+        <AnalyticsLoader />
       </head>
       <body className="flex min-h-screen flex-col">
         {/* Accessible skip link */}
