+++ package.json
@@ -63,6 +63,6 @@
     ]
   },
   "engines": {
-    "node": ">=18 <21"
+    "node": ">=18"
   }
 }
